test(redis): run the claim tests against Redis Cluster too
The claim tests become a `StreamClaimTestcase` shared by the single-node
and cluster configs, following the base-testcase model. The only
client-specific point is which class the unsupported-server test patches,
exposed as `client_cls`.
